Move parsing code out of common for namespaces and into pkg/parse.go
podman needs this code to be vendored in for podman build.
